Fixed all bugs

+++ Projecta-Backend/src/main/java/com/hackademics/repository/SubjectRepository.java
@@ -1,13 +1,13 @@
 package com.hackademics.repository;
 
-import com.hackademics.Model.Subject;
+import com.hackademics.model.Subject;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface SubjectRepository extends JpaRepository<Subject, Integer> {
-    Optional<Subject> findBySubjectId(int subjectId);
+public interface SubjectRepository extends JpaRepository<Subject, Long> {
+    Optional<Subject> findBySubjectId(Long subjectId);
     Optional<Subject> findBySubjectName(String subjectName);
 }
